backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic[email]==2.9.2
pydantic-settings==2.6.1
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.12
slowapi==0.1.9

# Testing
pytest==8.3.3
httpx==0.27.2
pytest-asyncio==0.24.0
